pyproject.toml
[project]
name = "desktop-agent"
version = "0.1.0"
description = "Agent S core logic refactored into three clean modules"
dependencies = [
    "pyautogui>=0.9.54",
    "Pillow>=10.0",
    "openai>=1.30",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["black", "ruff", "pytest"]
anthropic = ["anthropic>=0.18"]
gemini = ["google-generativeai>=0.3"]
all = [
    "anthropic>=0.18",
    "google-generativeai>=0.3",
]

[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"